Refactoring loop
I refactored my loop to read the lines in sequence: fullNameLine, numClassesLine, and then studentIdLine. I validated the data using the methods I created. Additionally, I included a check to see if the fullNameline is empty due to an input mistake. In such cases, it will skip the next two lines to prevent any issues in my code and resume reading from the fourth line onward.
After the validations are completed and return true, the data will be processed and written using my writer.